Support multiple RPMTAG_I18NSTRING arrays
While LSB specifies there should never be more than one I18NSTRING
per header, this is not defined in the RPM spec, nor is it implmented
in Redhat's RPM libraries.

To maintain compatability, for headers which have only 1 I18NSTRING
the header value is returned as a string.  For headers which have
more than 1 I18NSTRING (and would previously have thrown an exception)
we now return a list of strings

Signed-Off-By:  Ben.Chalmers@citrix.com